Première passe pour prendre en compte les specs SGF.
* Tout reste à faire, ce n'est que le début d'un long périple
* le board stocke des byte au lieu des int

+++ src/org/beroot/android/goengine/StringData.java
@@ -13,7 +13,7 @@ public class StringData
   /**
    * Color of string, BLACK or WHITE
    */
-  public int color;
+  public byte color;
 
   /**
    * Number of stones in string.
@@ -65,7 +65,7 @@ public class StringData
    * @param pos position initiale de la chaine
    * @param color couleur de la chaine
    */
-  public StringData(int boardSize, int pos, int color)
+  public StringData(int boardSize, int pos, byte color)
   {
     // Initialisation des structures
     stones = new int[boardSize];
